Change id data type to long and create service and dto for master

+++ src/main/java/com/app/smartdrive/api/entities/master/CarModel.java
@@ -17,13 +17,13 @@ public class CarModel {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "carm_id", updatable = false, nullable = false)
-    private int carmId;
+    private Long carmId;
 
     @Column(name = "carm_name", unique = true, nullable = false)
     private String carmName;
 
     @Column(name = "carm_carb_id", nullable = false)
-    private int carmCarbId;
+    private Long carmCarbId;
 
     @ManyToOne
     @JoinColumn(name = "carm_carb_id", insertable = false, updatable = false)
